fix(auth): 인증 코드 보낸 기록이 있는 경우, 삭제하고 코드 저장하도록 로직 변경

+++ gdgoc/src/main/java/inha/gdgoc/domain/auth/service/AuthCodeService.java
@@ -19,7 +19,9 @@ public class AuthCodeService {
 
     @Transactional
     public void saveAuthCode(String email, String code) {
-        authCodeRepository.deleteByEmail(email);
+        if(authCodeRepository.existsByEmail(email)) {
+            authCodeRepository.deleteByEmail(email);
+        }
         authCodeRepository.save(new AuthCode(email, code));
     }
